spr36279: Fix space-req computation for hlist-panes in windows.
hlist-panes are similar to text-editors in that they have a scroller-pane as 
a parent, but they provide their own scroll bars. (This is for layout 
reasons, apparently.) This change updates hlist-panes' compose-space methods 
to correctly compute the space requirements for their items, so that the 
vertical scroll bar doesn't overlap longer item names when it doesn't have to.