spr37149: Plug leak of the #\{ dispatch character.
This change eval-whens out the definition of the #\{ macro character
so that it happens only at compile time. The definition was leaked
into customer environments, and affected the system readtable when
loading CLIM from a clinit.cl file, which resulted in an error.

Tested by building & loading the resulting clim fasls on windows &
linux.