bug19469: Reset mod state on window activation.
This affects Windows only. When cycling through windows using Alt-Tab,
cycling back to the CLIM application would cause the Alt modifier state
to still be set, which disabled all input and mouse processing - the app
would appear to be hung.

This patch resets the port modifier state on window activation, which
allows CLIM to notice that Alt is no longer depressed.

Tested by customer and by me on Windows 7.