bug20152: Improve error message if libXm fails to load on mac
Make CLIM collect foreign load errors if a required .dylib file is
present but fails to load, and show them to the user who will
hopefully be able to diagnose what's wrong (or at the least, send us a
note about the error).

Is bootstrapping needed?   no
Are tests included for new features?
   no new features
Tests run:
   Tested manually on bigmac, and by Mike Travers (SRI)

Change-Id: Ida3a844d33dd8f337d6f2cf1f2f943d4f2033ccd